hero_list in deathfinder
Started a better way than regexes to extract data from the information - make a list of lists with each line from the index, and go through that smart-ly. Expunge any entry that isn't the first or 'title' or 'killed by' and you'll have everything you need in a neat bundle much easier.